[TOOLS] Fix/suppress all MSVC/x64 warnings

+++ sdk/lib/cmlib/hivebin.c
@@ -16,7 +16,7 @@ HvpAddBin(
 {
     PHMAP_ENTRY BlockList;
     PHBIN Bin;
-    SIZE_T BinSize;
+    ULONG BinSize;
     ULONG i;
     ULONG BitmapSize;
     ULONG BlockCount;
